FE-Nithi-014: Fix transaction details
